pyproject.toml
[project]
name = "copper"
version = "0.1.0"
description = "Mentecobres: AI-maintained knowledge bases, inspired by Karpathy and the Cosmere"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.12",
    "rich>=13.0",
    "pyyaml>=6.0.3",
    "httpx>=0.25",
    "python-dotenv>=1.2.2",
    "fastapi>=0.135.3",
    "uvicorn[standard]>=0.44.0",
    "python-multipart>=0.0.26",
    "jinja2>=3.1.6",
    "sse-starlette>=3.3.4",
    "loguru>=0.7.3",
    "pydantic-settings>=2.13.1",
    "core-utils @ git+https://github.com/fermaat/core-utils.git",
]

[project.scripts]
copper = "copper.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
    "isort>=5.12",
]
llm = [
    "core-llm-bridge @ git+https://github.com/fermaat/core-llm-bridge.git",
]
pdf = [
    "pdfplumber>=0.11",
]
watch = [
    "watchdog>=4.0",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
    "isort>=5.12",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

# Tests, Typer-based CLI commands, and FastAPI routes intentionally skip return
# annotations (decorators drive return shape).
[[tool.mypy.overrides]]
module = [
    "tests.*",
    "copper.cli",
    "copper.api.app",
    "copper.api.routes.*",
    "copper.llm.*",
    "copper.ingest.pdf",
    "copper.workflows.polish",
]
disallow_untyped_defs = false

# External libraries without type stubs.
[[tool.mypy.overrides]]
module = [
    "yaml",
    "requests",
    "pdfplumber",
    "core_llm_bridge.*",
    "watchdog.*",
]
ignore_missing_imports = true

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
ignore = ["E501", "E203"]
